+++ packages/terra-compact-interactive-list/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 1.10.0 - (March 14, 2024)
+
+* Added
+  * Added `hasVisibleBorders` prop to hide visual borders if needed.
+
+* Changed
+  * Updated to support `terra-menu` component as a cell content.
+
 ## 1.9.0 - (March 8, 2024)
 
 * Changed