+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+* Removed
+  * `boundingRef` is no longer necessary for the section header's sticky behavior for tables. This change is still passive in nature as it doesn't change any functionality for consumers.
+
+## 5.13.0 - (March 14, 2024)
+
+* Added
+  * Added `event` object to `onCellSelect` callback.
+
+* Changed
+  * Internal focus trap update for terra-data-grid.
+
 ## 5.12.0 - (March 8, 2024)
 
 * Changed